async: add doc
